% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{bilinear_binary_strings}
\alias{bilinear_binary_strings}
\title{Produce a set of n character strings of length nstr reflecting two distinct pathways
"leftwards" and "rightwards" through the bitstring}
\usage{
bilinear_binary_strings(n, nstr)
}
\arguments{
\item{n}{numeric number of strings to produce}

\item{nstr}{numeric length of binary strings}
}
\value{
character vector set of strings
}
\description{
Produce a set of n character strings of length nstr reflecting two distinct pathways
"leftwards" and "rightwards" through the bitstring
}
\examples{
bilinear_binary_strings(10, 4)
}
