% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{binary_strings_with_k_ones}
\alias{binary_strings_with_k_ones}
\title{Produce the set of binary strings of length nstr with k 1's}
\usage{
binary_strings_with_k_ones(nstr, k)
}
\arguments{
\item{nstr}{numeric length of binary strings}

\item{k}{numeric number of 1s in the strings}
}
\value{
character vector set of strings
}
\description{
Produce the set of binary strings of length nstr with k 1's
}
\examples{
binary_strings_with_k_ones(5, 3)
}
