% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{BinToDec}
\alias{BinToDec}
\title{Binary to decimal helper function}
\usage{
BinToDec(x)
}
\arguments{
\item{x}{string binary value to convert}
}
\value{
numeric decimal value
}
\description{
Binary to decimal helper function
}
\examples{
BinToDec("1001")
}
