% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{branching_count}
\alias{branching_count}
\title{Compute excess branching score of a graph}
\usage{
branching_count(g)
}
\arguments{
\item{g}{graph to analyse}
}
\value{
numeric excess branching score
}
\description{
Compute excess branching score of a graph
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
branching_count(sol$best.graph)
}
