% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{DecToBin}
\alias{DecToBin}
\title{Decimal to binary helper function}
\usage{
DecToBin(x, len)
}
\arguments{
\item{x}{numeric value to convert}

\item{len}{numeric length of required binary string}
}
\value{
string binary value
}
\description{
Decimal to binary helper function
}
\examples{
DecToBin(9, 4)
}
