% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{DecToBinV}
\alias{DecToBinV}
\title{Decimal to (numeric) binary helper function}
\usage{
DecToBinV(x, len)
}
\arguments{
\item{x}{numeric value to convert}

\item{len}{numeric length of required binary string}
}
\value{
numeric binary vector
}
\description{
Decimal to (numeric) binary helper function
}
\examples{
DecToBinV(9, 4)
}
