% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{is_compat}
\alias{is_compat}
\title{Check compatibility between two bitstrings, reported as a distance}
\usage{
is_compat(s1, s2)
}
\arguments{
\item{s1}{character bit string 1}

\item{s2}{character bit string 2}
}
\value{
numeric difference
}
\description{
Check compatibility between two bitstrings, reported as a distance
}
\examples{
is_compat("101", "001")
}
