% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is_compat_num}
\alias{is_compat_num}
\title{Check compatibility between two numeric bitstrings, reported as a distance}
\usage{
is_compat_num(s1, s2)
}
\arguments{
\item{s1}{numeric bit vector 1}

\item{s2}{numeric bit vector 2}
}
\value{
numeric difference
}
\description{
Check compatibility between two numeric bitstrings, reported as a distance
}
\examples{
is_compat_num(c(1,0,1), c(0,0,1))
}
