% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{layer_sum}
\alias{layer_sum}
\title{Compute layer sum of a graph}
\usage{
layer_sum(g)
}
\arguments{
\item{g}{graph to analyse}
}
\value{
numeric layer sum
}
\description{
Compute layer sum of a graph
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
layer_sum(sol$best.graph)
}
