% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{matdiff}
\alias{matdiff}
\title{Difference helper function for vectorised calculations}
\usage{
matdiff(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
numeric difference
}
\description{
Difference helper function for vectorised calculations
}
\examples{
matdiff(5, 2)
}
