% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{plot_stage_1}
\alias{plot_stage_1}
\title{Plot the unrewired arboresence (output from Gutin algorithm) and the rewired version increasing layer sum}
\usage{
plot_stage_1(graphs)
}
\arguments{
\item{graphs}{list output of model fit}
}
\value{
ggarrange object containing plots
}
\description{
Plot the unrewired arboresence (output from Gutin algorithm) and the rewired version increasing layer sum
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
plot_stage_1(sol)
}
