% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-functions.R
\name{plot_stage_2}
\alias{plot_stage_2}
\title{Plot the simplest spanning arborescence and simplest transition-spanning DAG}
\usage{
plot_stage_2(graphs)
}
\arguments{
\item{graphs}{list output of model fit}
}
\value{
ggarrange object containing plots
}
\description{
Plot the simplest spanning arborescence and simplest transition-spanning DAG
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
plot_stage_2(sol)
}
