% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-functions.R
\name{plot_stage_p}
\alias{plot_stage_p}
\title{Plot a solution graph}
\usage{
plot_stage_p(graphD, v.labels = data.frame(Species = NA))
}
\arguments{
\item{graphD}{graph to plot}

\item{v.labels}{data frame of vertex labels (optional)}
}
\value{
ggarrange object containing plots
}
\description{
Plot a solution graph
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
plot_stage_p(sol$best.graph)
}
