% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{plot_tree_data}
\alias{plot_tree_data}
\title{Plot a phylogeny and associated set of bitstrings at the tips}
\usage{
plot_tree_data(my.tree, tip.data)
}
\arguments{
\item{my.tree}{tree object}

\item{tip.data}{matrix or data frame containing one row for each tip}
}
\value{
ggplot object
}
\description{
Plot a phylogeny and associated set of bitstrings at the tips
}
\examples{
#my.tree = star_tree(4)
#tip.data = as.list(matrix(0, nrow=4, ncol=5))
#plot_tree_data(my.tree, tip.data)
print(999)
}
