% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot a solution graph with edges weighted by downstream datapoints}
\usage{
plot_weights(graphD, labels = c(""), thresh = 5)
}
\arguments{
\item{graphD}{graph to plot}

\item{labels}{vector of vertex labels (optional)}

\item{thresh}{numeric threshold below which edges are not plotted (default 5)}
}
\value{
ggarrange object containing plots
}
\description{
Plot a solution graph with edges weighted by downstream datapoints
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
plot_weights(sol$best.graph, thresh = 0)
}
