% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{random_binary_strings}
\alias{random_binary_strings}
\title{Produce a set of n random binary strings of length nstr}
\usage{
random_binary_strings(n, nstr)
}
\arguments{
\item{n}{numeric number of strings to produce}

\item{nstr}{numeric length of binary strings}
}
\value{
character vector set of strings
}
\description{
Produce a set of n random binary strings of length nstr
}
\examples{
random_binary_strings(5, 3)
}
