% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{simplest_arborescence}
\alias{simplest_arborescence}
\title{Compute the simplest spanning arborescence for a dataset}
\usage{
simplest_arborescence(ancnames, descnames = NULL)
}
\arguments{
\item{ancnames}{character vector of states}

\item{descnames}{character vector of descendant states (optional)}
}
\value{
list: names (state list), raw.graph (first solution), rewired.graph (minimised layer sum), raw.bc (branching count of first solution), rewired.bc (branching count of rewired solution), raw.ls (layer sum of first solution), rewired.ls (layer sum of rewired solution)
}
\description{
Compute the simplest spanning arborescence for a dataset
}
\examples{
simplest_arborescence(c("1011", "1001", "0100"))
}
