% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{simplest_DAG}
\alias{simplest_DAG}
\title{Compute the simplest transition-spanning DAG for ancestor-descendant data}
\usage{
simplest_DAG(ancnames, descnames)
}
\arguments{
\item{ancnames}{character vector of states}

\item{descnames}{character vector of descendant states}
}
\value{
list: names (state list), raw.graph (first solution), rewired.graph (minimised layer sum), raw.bc (branching count of first solution), rewired.bc (branching count of rewired solution), raw.ls (layer sum of first solution), rewired.ls (layer sum of rewired solution), dataset (data), best.graph (best solution), best.bc (branching count of best solution)
}
\description{
Compute the simplest transition-spanning DAG for ancestor-descendant data
}
\examples{
simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
}
