% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{simulate_accumulation}
\alias{simulate_accumulation}
\title{Simulate dynamics for L characters on a given tree, or a simulated tree with n tips}
\usage{
simulate_accumulation(
  tree.size,
  L,
  tree.type = "random",
  birth.rate = 1,
  death.rate = 0.5,
  accumulation.rate = 1,
  dynamics = "random",
  use.tree = NULL,
  char.probs = NULL
)
}
\arguments{
\item{tree.size}{numeric number of tips on tree. If 0, we assume that a tree is being provided in use.tree}

\item{L}{numeric number of characters to simulate}

\item{tree.type}{character type of tree to simulate. "random" is only option for now}

\item{birth.rate}{numeric birth rate for birth-death process simulating random tree}

\item{death.rate}{numeric death rate for birth-death process simulating random tree}

\item{accumulation.rate}{numeric rate of accumulation of features}

\item{dynamics}{character type of accumulation dynamics to simulate. "random" (all characters random, max one change per branch); "linear" (single pathway dependence); "poisson" (all characters random, Poisson accumulation); "mixed" (mix of linear and random); "heterogeneous" (random, with different relative probabilities in char.probs). default "random"}

\item{use.tree}{NULL or tree object. if tree.size is 0, the tree provided here will be used instead of simulating a new one}

\item{char.probs}{numeric vector of relative probabilities of each character acquisition, for "heterogeneous" accumulation}
}
\value{
a list: x (data at nodes and tips); my.tree (tree structure); ancnames (labels of ancestors for each transition); descnames (labels of descendants for each transition)
}
\description{
Simulate dynamics for L characters on a given tree, or a simulated tree with n tips
}
\examples{
simulate_accumulation(32, 4)
}
