% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{simulate_null_model}
\alias{simulate_null_model}
\title{Simulate from a null model of no character interactions, given a dataset}
\usage{
simulate_null_model(m, acc.rate = 1, tree = NULL)
}
\arguments{
\item{m}{vector of data}

\item{acc.rate}{numeric (default 1). For phylogenetic data, the null model involves simulation on the given tree, with character acquisition rates proportional to each character's prevalence in the data. This param gives the constant of proportionality. Choose it so that the number of positive observations in the null simulation is comparable to that in the original data.}

\item{tree}{tree (optional). Tree linking the observations.}
}
\value{
a named list, containing the solution object for this null model instance, and a dataframe with fit_properties of the solution
}
\description{
Simulate from a null model of no character interactions, given a dataset
}
