% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{star_tree}
\alias{star_tree}
\title{Create a star phylogeny with n tips}
\usage{
star_tree(n)
}
\arguments{
\item{n}{numeric number of tips}
}
\value{
tree object
}
\description{
Create a star phylogeny with n tips
}
\examples{
star_tree(4)
}
