% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-statistics.R
\name{transitions_spanned}
\alias{transitions_spanned}
\title{Verify that a solution graph spans the required transitions}
\usage{
transitions_spanned(g, ancnames, descnames)
}
\arguments{
\item{g}{graph of fitted model}

\item{ancnames}{vector of ancestor states}

\item{descnames}{vector of descendant states}
}
\value{
Boolean: are edges spanned?
}
\description{
Verify that a solution graph spans the required transitions
}
\examples{
ancnames = c("000", "010", "100")
descnames = c("110", "110", "101")
sol = simplest_DAG(ancnames, descnames)
transitions_spanned(sol$best.graph, ancnames, descnames)
}
