% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-algorithm.R
\name{write_single_steps}
\alias{write_single_steps}
\title{Write a solution to a file, choosing a particular set of single steps for each multistep change (used to compare HyperHMM output)}
\usage{
write_single_steps(trans, L, fname)
}
\arguments{
\item{trans}{XXX}

\item{L}{XXX}

\item{fname}{string filename}
}
\value{
XXX
}
\description{
Write a solution to a file, choosing a particular set of single steps for each multistep change (used to compare HyperHMM output)
}
\examples{
#sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
#write_single_steps(sol)
print(999)
}
