% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-statistics.R
\name{fit_properties}
\alias{fit_properties}
\title{Summarise a set of properties of a fitted HyperDAGs model}
\usage{
fit_properties(fit, verbose = FALSE)
}
\arguments{
\item{fit}{model structure to analyse}

\item{verbose}{Boolean, whether to output summary as a message (default false)}
}
\value{
data frame of summary properties
}
\description{
Summarise a set of properties of a fitted HyperDAGs model
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
fit_properties(sol)
}
