% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-functions.R
\name{plot_stage_gen}
\alias{plot_stage_gen}
\title{Plot a solution graph with control over styling}
\usage{
plot_stage_gen(
  graphD,
  v.labels = data.frame(Species = NA),
  label.size = 0,
  point.size = 0,
  edge.alpha = 1,
  label.style = 0
)
}
\arguments{
\item{graphD}{graph to plot}

\item{v.labels}{data frame of vertex labels (optional)}

\item{label.size}{numeric node label size (default chooses according to length)}

\item{point.size}{numeric node point size (default 0)}

\item{edge.alpha}{numeric edge transparency (default 1)}

\item{label.style}{character node plot style ("full" binary strings, "points" just points, or "labels" individual codes for nodes, default chooses according to length)}
}
\value{
ggarrange object containing plots
}
\description{
Plot a solution graph with control over styling
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
plot_stage_gen(sol$best.graph)
}
