% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-functions.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot a solution graph with edges weighted by downstream datapoints}
\usage{
plot_weights(
  graphD,
  labels = c(""),
  thresh = 5,
  edge.label.size = 3,
  check.overlaps = FALSE,
  orient = "across"
)
}
\arguments{
\item{graphD}{graph to plot}

\item{labels}{vector of vertex labels (optional)}

\item{thresh}{numeric threshold below which edges are not plotted (default 5)}

\item{edge.label.size}{numeric edge label size (default 3)}

\item{check.overlaps}{Boolean, check for label overlaps (default FALSE)}

\item{orient}{string, edge label orientation "across", "along" (default "across")}
}
\value{
list object containing a collection of plots: full and thresholded plots, and full and thresholded graphs
}
\description{
Plot a solution graph with edges weighted by downstream datapoints
}
\examples{
sol = simplest_DAG(c("000", "010", "100"), c("110", "110", "101"))
plot_weights(sol$best.graph, thresh = 0)
}
